+++ CHANGELOG.md
@@ -17,7 +17,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.4.0] - Unreleased
 
 ### Added
-- Method: `OxidEsales\Codeception\Module\OxideshopModules::uninstallModule`
+- Method:
+    - `OxidEsales\Codeception\Module\OxideshopModules::uninstallModule`
+    - `OxidEsales\Codeception\Module\Database::grabConfigValueFromDatabase`
+    - `OxidEsales\Codeception\Module\Oxideshop::regenerateDatabaseViews`
 
 ## [1.3.0] - 2020-07-06